move to helper class
