+++ starsky/starskytest/Services/ReplaceServiceTest.cs
@@ -38,7 +38,7 @@ namespace starskytest.Services
 			_iStorage = new FakeIStorage(new List<string>{"/"}, 
 				new List<string>{"/test.jpg","/test2.jpg", "/readonly/test.jpg", "/test.dng"});
 			_metaReplace = new MetaReplaceService(_query,new AppSettings{ ReadOnlyFolders = new List<string>{"/readonly"}},
-				new FakeSelectorStorage(_iStorage));
+				new FakeSelectorStorage(_iStorage), new FakeIWebLogger());
 
 		}
